web-server/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
dbfread==2.0.7
defusedxml==0.7.1
Django==5.0.6
et-xmlfile==1.1.0
ezodf==0.3.2
gunicorn==22.0.0
idna==3.7
lxml==5.2.2
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
psutil==5.9.8
psycopg2-binary==2.9.9
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
